[TASK] Use country short name instead of official name by default
It's easier for FE Users to find their country if the list displays the countries short name localized instead of official name. Eg.:
"United Kingdom" instead of "United Kingdom of Great Britain and Northern"
"Deutschland" instead of "Bundesrepublik Deutschland"
"Italia" instead of "Repubblica Italiana"

Also, it's easier for a BE editor to find a country in a long list if countries are displayed with their short name instead of official name.